+++ src/main/java/idealist/dao/handlers/columns/ShortColumnHandler.java
@@ -16,6 +16,6 @@ public class ShortColumnHandler implements ColumnHandler {
 
     @Override
     public Object apply(ResultSet rs, int columnIndex) throws SQLException {
-        return Short.valueOf(rs.getShort(columnIndex));
+        return rs.getShort(columnIndex);
     }
 }
